Tests/LibWeb/Text/expected/async-scrolling/non-scrollable-wheel-listener-prevents-viewport-scroll.txt

wheel routing admission: accepted
wheel scroll admission over map: blocked-by-wheel-event-region
window scrollY before wheel: 900
wheel event count: 1
wheel event cancelable: true
wheel event defaultPrevented: true
window scrollY: 900
